% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_prevEst.R
\name{adjust_prevEst}
\alias{adjust_prevEst}
\title{Create crude and adjusted rates from prevEst output}
\usage{
adjust_prevEst(
  prevEst_ouput = prevest,
  adjust = F,
  grouped_ages = T,
  groups = seq(0, 100, 5),
  max.age = 85,
  census.population = census.population,
  standard.population = US_2000_standard_population
)
}
\arguments{
\item{prevEst_ouput}{output from prevEst}

\item{adjust}{Logical. If TRUE, rates will be age-adjusted}

\item{grouped_ages}{Logical. If TRUE, multi-year age groups are used for adjustment.}

\item{groups}{If grouped ages are used, vector of grouped age levels contain the lowest group in each age group is selected}

\item{census.population}{description}

\item{standard_population}{description}
}
\value{
A dataframe with prevalence rates by age at prevalence
}
\description{
The function generates adjusted rates from prevalence output
}
