% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Canada_2011_standard_population}
\alias{Canada_2011_standard_population}
\title{2011 Canadian standard population in single ages from 0-90}
\format{
A data frame with 91 rows and 2 columns:
\describe{
\item{age}{Age}
\item{pop}{Population}
...
}
}
\description{
2011 Canadian standard population in single ages from 0-90
}
\keyword{datasets}
