% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census.population}
\alias{census.population}
\title{Census data from 2018
#'}
\format{
A data frame with 86 rows and 2 columns:
\describe{
\item{age}{Age in 2018}
\item{pop}{Total population}
...
}
}
\description{
Census data from 2018
#'
}
\keyword{datasets}
