% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_incidence_estimates.R
\name{compare_incidence_estimates}
\alias{compare_incidence_estimates}
\title{Visualize incidence estimates from multiple datasets}
\usage{
compare_incidence_estimates(
  incidence1 = NULL,
  incidence2 = NULL,
  complete_incidence = NULL,
  projected_incidence = NULL,
  names = c(incidence1 = "incidence1", incidence2 = "incidence2", projected_incidence =
    "Projected incidence", complete_incidence = "Complete incidence")
)
}
\arguments{
\item{incidence1}{Incidence dataframe for primary incidence data set (required).}

\item{incidence2}{Incidence dataframe for dataset used to estimate complete incidence (optional).}

\item{complete_incidence}{Incidence dataframe for complete incidence estimated using both datasets (optional).}

\item{projected_incidence}{Incidence dataframe for projected incidence (optional).}

\item{names}{Named character vector}
}
\description{
Generates a visual comparison of multiple incidence estimates by year. Requires properly formated incidence estimates.
}
\details{
This function requires at least one incidence estimate. It will generate a ggplot comparing up to 4 incidence estimates from a main data set, comparison dataset, complete incidence, and project incidence files
}
\examples{

 \dontrun{
data(incidence)
incidence1 <- format_incidence(data=incidence,
                  ages = 0:85,
                  years = c(2010:2018),
                  names = c("ageDiag" = "ageDiag",
                          "yrDiag" = "yrDiag",
                          "incidence" = "count"),
                           keepExtraCols=FALSE)
                           
data(incidence_est)
incidence2 <- format_incidence(data=incidence_est,
                  ages = 0:85,
                  years = c(2010:2018),
                  names = c("ageDiag" = "ageDiag",
                          "yrDiag" = "yrDiag",
                          "incidence" = "count"),
                           keepExtraCols=FALSE)
                           
complete_incidence <- regPrev(incidence=incidence1,
                  incidence_est=incidence2,
                  regYr=2004:2021,
                  durationYr=1975:2009 )
                  
projected_incidence <- project_incidence(data=complete_incidence,
                  projection.years=2022:2025)                   

compare_incidence_estimates(incidence1=incidence1,
                  incidence2=incidence2,
                  complete_incidence=complete_incidence,
                  projected_incidence=projected_incidence,
                  names=c("incidence1"="incidence1",
                          "incidence2"="incidence2",
                          "projected_incidence"="Projected incidence",
                          "complete_incidence"="Complete incidence")) 
}
}
