% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_prevalence_estimates.R
\name{compare_prevalence_estimates}
\alias{compare_prevalence_estimates}
\title{Visualize prevalence estimates from multiple methods}
\usage{
compare_prevalence_estimates(
  prevalence1 = NULL,
  prevalence2 = NULL,
  prevalence3 = NULL,
  prevalence4 = NULL,
  names = c(prevalence1 = "prevalence1", prevalence2 = "prevalence2", prevalence3 =
    "prevalence3", prevalence4 = "prevalence4"),
  include_counts = FALSE
)
}
\arguments{
\item{prevalence1}{Dataframe of prevalence estimates (required).}

\item{prevalence2}{Dataframe of prevalence estimates (optional).}

\item{prevalence3}{Dataframe of prevalence estimates (optional).}

\item{prevalence4}{Dataframe of prevalence estimates (optional).}

\item{names}{Named character vector}

\item{include_counts}{Logical. Will print text of prevalence counts above bars.}
}
\description{
Generates a visual comparison of multiple prevalence estimates.
}
\details{
This function requires at least one prevalence estimate. It will generate a ggplot comparing up to 4 prevalence estimates.
}
