% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{fill_survival}
\alias{fill_survival}
\title{Apply survival stable after observation period}
\usage{
fill_survival(survival.data = survival.data)
}
\arguments{
\item{survival.data}{Survival dataframe}
}
\description{
This function works within a survival formatting function to make survival stable after end of observation period
}
