% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_incidence.R
\name{format_incidence}
\alias{format_incidence}
\title{Format incidence data}
\usage{
format_incidence(
  data,
  ages,
  years,
  names = c(ageDiag = "age", yrDiag = "year", incidence = "count"),
  keepExtraCols = FALSE
)
}
\arguments{
\item{data}{Incidence dataframe}

\item{ages}{Numeric vector}

\item{years}{Numeric vector}

\item{names}{Named character vector}

\item{keepExtraCols}{Logical}
}
\value{
A formatted incidence dataframe.
}
\description{
This function formats incidence data for ease of use with functions included in this package. It takes new cases and summarizes them by age and year of diagnosis.
}
\details{
The prevEst function requires properly formatted incidence and survival data. This function, the counterpart to \code{\link[=format_survival]{format_survival()}}, is designed
to take SEER-like incidence data and format it for the \code{\link[=prevEst]{prevEst()}} function. Three columns are necessary: 1) age at diagnosis, 2) year of diagnosis,
and 3) the reported incidence for that combination of the two. While these functions are not mandatory, they help prepare the data for \code{\link[=prevEst]{prevEst()}} with simple transformation steps.
}
\examples{

data(incidence)
format_incidence(incidence,
                  ages = seq(0, 85, 5),
                  years = c(2010:2018),
                  names = c("ageDiag" = "ageDiag",
                          "yrDiag" = "yrDiag",
                          "incidence" = "count"))


}
\seealso{
\code{\link[=format_survival]{format_survival()}} The analogous function that formats survival data
}
