% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_lifetable.R
\name{format_lifetable}
\alias{format_lifetable}
\title{Format life table data}
\usage{
format_lifetable(
  data,
  names = c(age = "age", year = "yr", interval = "period", Expected =
    "Expected_Interval")
)
}
\arguments{
\item{data}{Incidence dataframe}

\item{names}{Named character vector}
}
\value{
A formatted life table dataframe.
}
\description{
This function formats life table data for ease of use with functions included in this package.
}
\details{
Survival projection functions in prevEst  requires properly formatted life table data. This function, the counterpart to \code{\link[=format_incidence]{format_incidence()}} and \code{\link[=format_survival]{format_survival()}}, is designed
to take SEER-like life table data and format it for the \code{\link[=prevEst]{prevEst()}} function. Four columns are necessary: 1) age, 2) year, 3) survival interval (in years),
and 4) expected interval survival. While these functions are not mandatory, they help prepare the data for \code{\link[=prevEst]{prevEst()}} with simple transformation steps.
}
\examples{

format_lifetable(life.table.raw,
                  names = c("age" = "age",
                          "year" = "yr",
                          "interval" = "period",
                          "Expected"="Expected_Interval"))


}
\seealso{
\code{\link[=format_incidence]{format_incidence()}} The analogous function that formats incidence data

\code{\link[=format_survival]{format_survival()}} The analogous function that formats survival data
}
