% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{format_survival}
\alias{format_survival}
\title{Format survival data}
\usage{
format_survival(
  data,
  SurvType = "Observed",
  ages = NULL,
  years = NULL,
  prevYear = NULL,
  names = c(ageDiag = "ageDiag", yrDiag = "yrDiag", yrPrev = "yrPrev", period = "period",
    Observed = "Observed", expected = "Expected")
)
}
\arguments{
\item{data}{Survival dataframe}

\item{SurvType}{Specify observed or expected survival}

\item{ages}{Numeric vector. If not included, defaults to all ages in data.}

\item{years}{Numeric vector. If not specified, defaults to all years in data.}

\item{prevYear}{Year of prevalence}

\item{names}{Named vector}
}
\value{
A formatted survival dataframe.
}
\description{
This function formats survival data for streamlined use with functions included in this package. While designed to work with case-listing, it can also work with already summarized survival data provided by the SEER*Stat "Survival Session".
}
\details{
The prevEst function requires properly formatted incidence and survival data. This function, the counterpart to \code{\link[=format_incidence]{format_incidence()}},is designed
to take SEER-like incidence data and format it for the \code{\link[=prevEst]{prevEst()}} function. Three columns are necessary: 1) age at diagnosis, 2) year of diagnosis,
and 3) the observed survival for the combination of the two. While these functions are not mandatory, they help prepare the data for \code{\link[=prevEst]{prevEst()}} with simple transformation steps.
}
\seealso{
\code{\link[=format_incidence]{format_incidence()}} The analogous function that formats incidence data
}
