% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incidence_est}
\alias{incidence_est}
\title{#' Simulated projected incidence data}
\format{
' A data frame with 1376 rows and 3 columns:
\describe{
\item{yrDiag}{Year of diagnosis}
\item{ageDiag}{Age at Diagnosis}
\item{count}{Number of cases}
...
}
}
\description{
Simulated cancer registry incidence data
}
\keyword{datasets}
