% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life.table}
\alias{life.table}
\title{Simulated life table data}
\format{
A data frame with 3,521 rows and 3 columns:
\describe{#'
\item{period}{years of survival}
\item{agePrev}{Age at prevalence}
\item{expected}{Percent expected to survive period}
...
}
}
\description{
Simulated cancer registry survival data
}
\keyword{datasets}
