% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{no_survival}
\alias{no_survival}
\title{Apply assumption of death after observation period}
\usage{
no_survival(
  survival.data = survival.data,
  prevYear = NULL,
  years.observed.surv = NULL
)
}
\arguments{
\item{survival.data}{Survival dataframe}

\item{prevYear}{Year for which prevalence is being estimated}

\item{years.observed.surv}{Years after which to go to population survival}
}
\description{
This function works within a survival formatting function to apply an assumption of death after observation period
}
