% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevEst.R
\name{prevEst}
\alias{prevEst}
\title{Estimate complete prevalence}
\usage{
prevEst(
  incidence,
  survival,
  year,
  years = NULL,
  adjust = F,
  grouped_ages = F,
  groups = NULL
)
}
\arguments{
\item{incidence}{Incomplete incidence dataframe containing age at diagnosis (named "ageDiag"), year of diagnosis (named "yrDiag"), and count of new cases (named "count)}

\item{survival}{Survival dataframe for each age at prevalence year containing:  age at diagnosis (named "ageDiag"), year of diagnosis (named "yrDiag"), observed survival rate (named "survival")}

\item{year}{Numeric. Year for which prevalence is to be estimated}

\item{years}{Numeric. Years to use for calculating prevalence. Defaults to all years included in incidence dataframe.}

\item{adjust}{Logical. If TRUE, rates will be age-adjusted}

\item{grouped_ages}{Logical. If TRUE, multi-year age groups are used for adjustment.}

\item{groups}{If grouped ages are used, vector of grouped age levels contain the lowest group in each age group is selected}
}
\value{
A dataframe with estimated prevalence by age at prevalence. If incidence data is grouped (as in the sample data),
then it will assume that all cases in the group (e.g. ages 0-4) have the same rates of survival.
}
\description{
The function \code{prevEst} is used to estimate complete prevalence using survival and incomplete incidence data
}
\examples{

data(incidence)
data(incidence_est)
data(survival)

\dontrun{
prevEst(incidence = regPrev(incidence = incidence,
                             incidence_est = incidence_est,
                            regYr = c(2001:2017)),
        survival = format_survival(survival,
                                   ages = c(0:85),
                                   years = c(1995:2018),
                                   assumption = "population",
                                   names = c("ageDiag" = "ageDiag",
                                             "yrDiag" = "yrDiag",
                                             "Observed" = "survival"),
                                   life.table = life.table),
        year = 2018,
        years = 2010:2018)         
}
}
