% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_incidence.R
\name{project_incidence}
\alias{project_incidence}
\title{Project incidence for unobserved year(s)}
\usage{
project_incidence(data, projection.years = NULL, method = "poisson")
}
\arguments{
\item{data}{Incidence dataframe for which prevalence is to be projected}

\item{projection.years}{Numeric vector of years to project prevalence for}

\item{method}{GLM family to be base predictions off of}
}
\value{
Predicted survival proportions for \code{years}

Projected incidence dataframe including projection years
}
\description{
Project incidence for unobserved year(s)
}
\examples{
\dontrun{
project_incidence(data=complete_incidence,
                  final.year=2030)
}
}
\seealso{
\code{\link[=project_survival]{project_survival()}} The analogous function that projects survival data
}
