% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{project_survival}
\alias{project_survival}
\title{Project survival data}
\usage{
project_survival(
  data,
  ages,
  observation.years = NULL,
  projection.years = NULL,
  prevYear = NULL,
  assumption = "nosurvival",
  life.table = NULL
)
}
\arguments{
\item{data}{Survival dataframe for which prevalence is to be projected}

\item{ages}{Numeric vector of ages to be included}

\item{observation.years}{Numeric vector of observed years to build the model on}

\item{projection.years}{Numeric vector of years to predict survival for}

\item{prevYear}{Year for which prevalence is being estimated}

\item{assumption}{Named character, either "population" or "nosurvival"}

\item{life.table}{Dataframe of SEER life table}

\item{names}{Named character vector pointing to column names}
}
\value{
Predicted survival including \code{projection.years}
}
\description{
Project survival data  for unobserved future or past year(s) requires properly formatted data
}
\examples{
\dontrun{
project_survival(data=raw.survival,
                 ages= 0:85,
                 observation.years = NULL,               # A vector of years to build the model on
                 projection.years = NULL,
                 prevYear = NULL,
                 assumption = "population",
                 life.table = life.tables,
                 names = c("ageDiag"="Age_recode_with_single_ages_and_85",
                           "yrPrev"="yrPrev",
                           "yrDiag"="Year_of_diagnosis",
                           "Observed"="Observed"),
                 keepExtraCols=FALSE)
}
}
\seealso{
\code{\link[=project_incidence]{project_incidence()}}
}
