% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regPrev.R
\name{regPrev}
\alias{regPrev}
\title{Estimate complete incidence}
\usage{
regPrev(incidence, incidence_est, regYr = NULL, durationYr = NULL)
}
\arguments{
\item{incidence}{Incomplete incidence dataframe containing age at diagnosis (named "ageDiag"), year of diagnosis (named "yrDiag"), and count of new cases (named "count)}

\item{incidence_est}{Estimated incidence dataframe for smaller population area (with longer follow-up) containing all years with age at diagnosis (named "ageDiag"), year of diagnosis (named "yrDiag"), and count of new cases (named "count)}

\item{regYr}{Years to do regression on i.e. c(2000:2010)}

\item{durationYr}{Years that require estimates from regression}
}
\value{
An incidence dataframe for the years provided.
}
\description{
Estimate complete incidence
}
\examples{

data(incidence)
data(incidence_est)

regPrev(incidence = incidence,
        incidence_est = incidence_est,
       regYr = c(2001:2017),
       durationYr = c(1975:2000))

}
