% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regSurv.R
\name{regSurv}
\alias{regSurv}
\title{Estimate complete survival}
\usage{
regSurv(
  surv,
  names = c(ageDiag = "ageDiag", yrDiag = "yrDiag", period = "period", observed =
    "survival"),
  life.table
)
}
\arguments{
\item{surv}{Incomplete survival dataframe containing age at diagnosis (named "ageDiag"), year of diagnosis (named "yrDiag"), and survival of new cases (named "observed)}

\item{names}{Years to do regression on i.e. c(2000:2010)}

\item{life.table}{Dataframe of life tables with relative survival with respect to age and period}
}
\value{
A survival dataframe with missing values imputed
}
\description{
Estimate complete survival
}
\examples{

\dontrun{
data(survival)
data(life.table)
regSurv(surv = survival,
        life.table = life.table)
}
}
