% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival}
\alias{survival}
\title{Simulated survival data (formatted)}
\format{
A data frame with 170,280 rows and 4 columns:
\describe{
\item{ageDiag}{Age at Diagnosis}
\item{period}{years of survival}
\item{yrDiag}{Year of diagnosis}
\item{survival}{Percent surviving}
...
}
}
\description{
Simulated cancer registry survival data
}
\keyword{datasets}
