% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival_unformatted}
\alias{survival_unformatted}
\title{Simulated survival data (unformated)}
\format{
A data frame with 170,280 rows and 4 columns:
\describe{
\item{ageDiag}{Age at Diagnosis}
\item{yrDiag}{Year of diagnosis}
\item{yrPrev}{Year at which survival was observed}
\item{Observed}{Percent surviving}
...
}
}
\description{
Simulated cancer registry survival data
}
\keyword{datasets}
