% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US_2000_standard_population}
\alias{US_2000_standard_population}
\title{2000 US standard population in single ages from 0-90}
\format{
A data frame with 91 rows and 2 columns:
\describe{
\item{age}{Age}
\item{pop}{Population}
...
}
}
\description{
2000 US standard population in single ages from 0-90
}
\keyword{datasets}
