% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{World_standard_population}
\alias{World_standard_population}
\title{World (WHO 2000-2025) Std Million from 0-90}
\format{
A data frame with 91 rows and 2 columns:
\describe{
\item{age}{Age}
\item{pop}{Population}
...
}
}
\description{
World (WHO 2000-2025) Std Million from 0-90
}
\keyword{datasets}
