% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_survival.R
\name{format_survival}
\alias{format_survival}
\title{Format survival data}
\usage{
format_survival(
  data,
  ages = NULL,
  years = NULL,
  prevYear = NULL,
  assumption = "nosurvival",
  life.table = NULL,
  names = c(ageDiag = "ageDiag", yrDiag = "yrDiag", Observed = "Observed"),
  keepExtraCols = FALSE
)
}
\arguments{
\item{data}{Dataframe}

\item{ages}{Numeric vector}

\item{years}{Numeric vector}

\item{prevYear}{Numeric vector}

\item{assumption}{Character vector}

\item{life.table}{Dataframe}

\item{names}{Named vector}

\item{keepExtraCols}{Logical}
}
\value{
Survival dataframe, formatted
}
\description{
This function formats survival data for easier/more streamlined use with functions included in this package. While designed to work with case-listing, it can also work with already summarized survival data provided by the SEER*Stat "Survival Session".
}
\details{
The prevEst function requires both a properly formatted incidence and survival data. This function, the counterpart to \code{\link[=format_incidence]{format_incidence()}}, is designed
to take SEER-like survival data and format it to work more easily with the \code{\link[=prevEst]{prevEst()}} function. 3 columns are necessary: a column for 1) age at diagnosis, 2) year of diagnosis,
and 3) the observed survival for that combination of the two. While these functions aren't necessary, they help wrap some simple transformation steps.
}
\examples{

format_survival(survDf,
                ages = c(0:85),
                years = c(2004:2018),
                assumption = "nosurvival",
                names = c("ageDiag" = "age",
                          "yrDiag" = "yrDiag",
                          "Observed" = "Observed"),
                keepExtraCols = FALSE)


format_survival(data = raw.survival,
                ages = 0:85,
                years = 2000:2018,
                prevYear = 2018,
                life.table = NULL,
                names = c("ageDiag" = "age",
                          "yrDiag" = "yrDiag",
                          "Observed" = "Observed")


}
\seealso{
\code{\link[=format_incidence]{format_incidence()}} The sister function that formats incidence data
}
