% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_survival.R
\name{format_survival}
\alias{format_survival}
\title{Format survival data}
\usage{
format_survival(
  data,
  ages = NULL,
  years = NULL,
  prevYear = NULL,
  assumption = "nosurvival",
  life.table = NULL,
  names = c(ageDiag = "ageDiag", yrDiag = "yrDiag", Observed = "Observed"),
  keepExtraCols = FALSE
)
}
\arguments{
\item{data}{Survival dataframe}

\item{ages}{Numeric vector}

\item{years}{Numeric vector}

\item{prevYear}{Numeric vector}

\item{assumption}{Named character vector}

\item{life.table}{Dataframe of SEER life table}

\item{names}{Named vector}

\item{keepExtraCols}{Logical}
}
\value{
A formatted survival dataframe.
}
\description{
This function formats survival data for streamlined use with functions included in this package. While designed to work with case-listing, it can also work with already summarized survival data provided by the SEER*Stat "Survival Session".
}
\details{
The prevEst function requires properly formatted incidence and survival data. This function, the counterpart to \code{\link[=format_incidence]{format_incidence()}},is designed
to take SEER-like incidence data and format it for the \code{\link[=prevEst]{prevEst()}} function. Three columns are necessary: 1) age at diagnosis, 2) year of diagnosis,
and 3) the observed survival for the combination of the two. While these functions are not mandatory, they help prepare the data for \code{\link[=prevEst]{prevEst()}} with simple transformation steps.
}
\examples{

data(survival)
data(life.table)

format_survival(survival,
                ages = c(0:85),
                years = c(2015:2018),
                assumption = "nosurvival",
                names = c("ageDiag" = "ageDiag",
                          "yrDiag" = "yrDiag",
                          "Observed" = "survival"),
                keepExtraCols = FALSE)
                
format_survival(survival,
                ages = c(0:85),
                years = c(1995:2018),
                assumption = "population",
                names = c("ageDiag" = "ageDiag",
                          "yrDiag" = "yrDiag",
                          "Observed" = "survival"),
                life.table = life.table,
                keepExtraCols = FALSE)




}
\seealso{
\code{\link[=format_incidence]{format_incidence()}} The analogous function that formats incidence data
}
