% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{no_survival}
\alias{no_survival}
\title{Apply assumption of death after observation period}
\usage{
no_survival(survival.data = survival.data, years.observed.surv = NULL)
}
\description{
This function works within a survival formatting function to apply an assumption of death after observation period
}
