% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{population_survival}
\alias{population_survival}
\title{Apply population survival assumption for years after observation}
\usage{
population_survival(
  life.table = life.table,
  survival.data = survival.data,
  prevYear = NULL,
  years.observed.surv = NULL
)
}
\description{
This function works within a survival formatting function to have survival go to population expected survival after end of observed survival ye
}
