% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevEst.R
\name{prevEst}
\alias{prevEst}
\title{Estimate complete prevalence}
\usage{
prevEst(
  incidence,
  survival,
  year,
  years = NULL,
  adjust = F,
  grouped_ages = F,
  groups = NULL,
  sex_specific = "Both sexes",
  incidence_est = NULL
)
}
\arguments{
\item{incidence}{Incomplete incidence data.frame containing age at diagnosis (named "ageDiag"), year diagnosed (named "yrDiag"),  count of new cases (named "count)}

\item{survival}{Survival data.frame for each age at prevalence year containing:  age at diagnosis (named "ageDiag"), year diagnosed (named "yrDiag"), observed survival rate (named "survival")}

\item{year}{Year for which prevalence is to be estimated}

\item{years}{Years to use for calculating prevalence. Defaults to all years included in incidence dataframe.}

\item{adjust}{logical. If TRUE, rates will be age adjusted}

\item{grouped_ages}{logical. If TRUE, multi-year age gruops are used for adjustment.}

\item{groups}{If grouped ages are use, vector of grouped age levels contain the lowest group in each agre groupe is selected}

\item{sex_specific}{character string containing sex ("Male" or "Female" if requesting sex-specific rates)}

\item{incidence_est}{Estimated incidence data frame for all years containing  age at diagnosis (named "ageDiag"), year diagnosed (named "yrDiag"),  count of new cases (named "count)}
}
\value{
Data.frame object
}
\description{
The function \code{prevEst} is used to estimate complete prevalence using survival and incomplete incidence data
}
\examples{

prevEst(incidence = incidence.reformat,
       survival = survival.nosurvival,
       year = 2018,
       adjust = T,
       standardpopulation = standard.population,
       censuspopulation = census.population,
       grouped_ages = T,
       groups = seq(0,85,5))


}
