% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_incidence.R
\name{project_incidence}
\alias{project_incidence}
\title{Project incidence for unobserved year}
\usage{
project_incidence(data, projection.years = NULL, method = "poisson")
}
\arguments{
\item{projection.years}{Incidence dataframe for which prevalence is to be projected}

\item{method}{GLM family to be base predictions off of}

\item{incidence}{Incidence dataframe for which prevalence is to be projected}
}
\value{
Predicted survival proportions for \code{years}

Projected incidence dataframe including projection years
}
\description{
Project incidence for unobserved year
}
\examples{
project_incidence(data=complete_incidence,
                  final.year=2030)
}
\seealso{
\code{\link[=project_survival]{project_survival()}}
}
