% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_incidence.R
\name{project_incidence}
\alias{project_incidence}
\title{Project incidence for unobserved year(s)}
\usage{
project_incidence(
  data,
  data.years = NULL,
  projection.years = NULL,
  method = "poisson"
)
}
\arguments{
\item{data}{Incidence dataframe for which incidence is to be projected}

\item{data.years}{years covered by incidence years data.frame (if not specified will use all years in the data)}

\item{projection.years}{Numeric vector of years to project incidence for}

\item{method}{GLM family to be base predictions off of}
}
\value{
Predicted survival proportions for \code{years}

Projected incidence dataframe including projection years
}
\description{
Project incidence for unobserved year(s)
}
\examples{
\dontrun{
project_incidence(data=complete_incidence,
data.years=NULL,
projection.years = NULL,
method = "poisson")
}
}
\seealso{
\code{\link[=project_survival]{project_survival()}} The analogous function that projects survival data
}
