% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_survival.R
\name{project_survival}
\alias{project_survival}
\title{Project survival data for use in prevEst}
\usage{
project_survival(
  data,
  ages,
  observation.years = NULL,
  projection.years = NULL,
  Observed.Surv = TRUE,
  Expected.Surv = FALSE,
  assumption = "nosurvival",
  life.table = NULL,
  names = c(ageDiag = "ageDiag", yrPrev = "yrPrev", yrDiag = "yrDiag", Observed =
    "observed"),
  keepExtraCols = FALSE
)
}
\arguments{
\item{data}{Survival dataframe for which prevalence is to be projected}

\item{ages}{Ages to be included}

\item{observation.years}{A vector of years to build the model on}

\item{projection.years}{A vector of years to predict survival for}

\item{Observed.Surv}{Logical}

\item{Expected.Surv}{Logical}

\item{assumption}{Character, either "population" or "nosurvival"}

\item{life.table}{SEER life table}

\item{names}{Named vector pointing to column names}

\item{keepExtraCols}{Logical}
}
\value{
Predicted survival including \code{projection.years}
}
\description{
Project survival data for use in prevEst
}
\examples{

project_survival(data=raw.survival,
                 ages= 0:85,
                 years=1975:2018,
                 final.year=2030,
                 observation.years = NULL,               # A vector of years to build the model on
                 projection.years = NULL,
                 assumption = "population",
                 life.table = life.tables,
                 years.observed.surv=20,
                 names = c("ageDiag"="Age_recode_with_single_ages_and_85",
                           "yrPrev"="yrPrev",
                           "yrDiag"="Year_of_diagnosis",
                           "Observed"="Observed"))

}
\seealso{
\code{\link[=project_incidence]{project_incidence()}}
}
