% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regPrev.R
\name{regPrev}
\alias{regPrev}
\title{Estimate complete prevalence}
\usage{
regPrev(incidence, incidence_est, regYr = NULL, durationYr = NULL)
}
\arguments{
\item{incidence}{Incomplete incidence data.frame containing age at diagnosis (named "ageDiag"), year diagnosed (named "yrDiag"),  count of new cases (named "count)}

\item{incidence_est}{Estimated incidence data frame for smaller area containing  all years containing  age at diagnosis (named "ageDiag"), year diagnosed (named "yrDiag"),  count of new cases (named "count)}

\item{regYr}{Years to do regression on ie. c(2000:2010)}

\item{durationYr}{Years that require estimates from regression}
}
\value{
An incidence dataframe
}
\description{
Estimate complete prevalence
}
\examples{
regprev(incidence = incomplete_incidence,
       regYr = c(2001:2017),
       durationYr = c(1975:2000))

}
